+++ LaClave/src/main/java/com/itwillbs/LaClave/review/ReviewRepository.java
@@ -6,17 +6,17 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface ReviewRepository extends JpaRepository<Review,Integer>{
+public interface ReviewRepository extends JpaRepository<Review,Long>{
 	
 	// 내가쓴 리뷰 목록
-	List<Review> findByMemberIdx(Integer memberIdx);
+	List<Review> findByMemberIdx(Long memberIdx);
 	
 	// 상품별 리뷰 목록 조회 
-    List<Review> findByProductIdxAndStatus(Integer productIdx, String status);
+    List<Review> findByProductIdxAndStatus(Long productIdx, String status);
     
     // 상품별 평균 점수 조회
     @Query("SELECT ROUND(AVG(r.score), 1) FROM Review r WHERE r.productIdx = :productIdx AND r.status = 'ACTIVE'")
-    Double getAverageScoreByProduct(@Param("productIdx") Integer productIdx);
+    Double getAverageScoreByProduct(@Param("productIdx") Long productIdx);
 
 	List<Review> findByProductIdx(int intValue);
 }
