+++ LaClave/src/main/java/com/itwillbs/LaClave/review/ReviewService.java
@@ -4,15 +4,15 @@ import java.util.List;
 
 public interface ReviewService {
 	
-	List<Review> getReviewBymember(Integer memberIdx);
+	List<Review> getReviewBymember(Long memberIdx);
 	//상품 별 리뷰 목록 조회
-	List<Review> getReviewByProduct(Integer productIdx,String status);
+	List<Review> getReviewByProduct(Long productIdx,String status);
 	
 	//상품 정보와 리뷰 조회
 	ReviewResponseDTO getProductWithReviews(Long productIdx);
 	
 	
-	Double getProductAverageScore(Integer productIdx);
+	Double getProductAverageScore(Long productIdx);
 	
 
 }
